requirements.txt
fastapi
python-decouple
uvicorn
pydantic
cloudinary
sqlalchemy
python-multipart
django
Mako
asyncpg
alembic
fastapi-utils
typing-inspect
pyjwt
passlib
bcrypt
netcat
pgsql